+++ src/plugins/opengl/src/painter_opengl.h
@@ -18,7 +18,7 @@ namespace Halley
 		void doStartRender() override;
 		void doEndRender() override;
 
-		void clear(std::optional<Colour> colour, std::optional<float> depth, std::optional<uint32_t> stencil) override;
+		void clear(std::optional<Colour> colour, std::optional<float> depth, std::optional<uint8_t> stencil) override;
 		void setMaterialPass(const Material& material, int pass) override;
 		void setMaterialData(const Material& material) override;
